requirements.txt
requests
numpy
pypdf
python-dotenv
groq
streamlit
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
torch
pytest